requirements.txt
glfw~=1.11.2
networkx~=2.4
tqdm~=4.19.4
trimesh~=3.7.4
scipy~=1.5.4
Dijkstar~=2.5.0
numpy~=1.19.5
Shapely~=1.7.0
matplotlib~=3.2.2
pandas~=1.0.5
gurobipy~=9.1.2
klampt~=0.8.7
Pillow~=8.2.0
PyOpenGL~=3.1.5
scikit_learn~=0.24.2
